finance-ai/backend/requirements.txt
fastapi
uvicorn
openai-whisper
torch
tqdm
numpy
pandas
scikit-learn
pytesseract
opencv-python
pillow
openai
tiktoken
python-multipart
requests
python-dotenv
langchain
faiss-cpu
pydub